move deprecated pop-up menu implementation into popup_menu.go
